// Project02/src/types/note.ts
export type Tag = {
  id: string
  label: string
}

export type NoteData = {
  title: string
  markdown: string
  tags: Tag[]
}

export type NoteType = {
  id: string
} & NoteData

export type RawNoteData = {
  title: string
  markdown: string
  tagIds: string[]
}

export type RawNote = {
  id: string
} & RawNoteData